refactor tool_handlers
