+++ skippy/assistants.go
@@ -2,65 +2,13 @@ package skippy
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log"
-	"strings"
 	"time"
 
 	openai "github.com/sashabaranov/go-openai"
 )
-
-const (
-	// ai functions
-	GetStockPriceKey     string = "get_stock_price"
-	GetWeatherKey        string = "get_weather"
-	SendChannelMessage   string = "send_channel_message"
-	SetReminder          string = "set_reminder"
-	GenerateImage        string = "generate_image"
-	ToggleMorningMessage string = "toggle_morning_message"
-	GenerateEvent        string = "generate_event"
-)
-
-type FuncArgs struct {
-	JsonValue string
-	FuncName  string
-	ToolID    string
-}
-
-type StockFuncArgs struct {
-	Symbol string
-}
-
-type WeatherFuncArgs struct {
-	Location string
-}
-
-type GenerateImageFuncArgs struct {
-	Prompt string `json:"prompt"`
-}
-
-type MorningMsgFuncArgs struct {
-	Enable           bool     `json:"enable"`
-	Time             string   `json:"time"`
-	WeatherLocations []string `json:"weather_locations,omitempty"`
-	Stocks           []string `json:"stocks,omitempty"`
-}
-
-// used for
-type ReminderFuncArgs struct {
-	Message     string `json:"message"`
-	TimerLength int    `json:"timer_length,omitempty"`
-	UserID      string `json:"user_id,omitempty"`
-}
-
-// TODO: find where this should live
-type EventFuncArgs struct {
-	Description         string `json:"description"`
-	Name                string `json:"name"`
-	NotificationMessage string `json:"notification_message"`
-}
 
 // TODO: update
 // TODO: these functions feel like they should exist on skippy struct, but I guess it doesn't matter
@@ -365,263 +313,6 @@ outerloop:
 		}
 	}
 	return submitToolOutputs(s.AIClient, toolOutputs, threadID, run.ID)
-}
-
-func handleGetWeather(funcArg FuncArgs, s *Skippy) (string, error) {
-	weatherFuncArgs := WeatherFuncArgs{}
-	err := json.Unmarshal([]byte(funcArg.JsonValue), &weatherFuncArgs)
-	if err != nil {
-		log.Println("Could not unmarshal func args: ", string(funcArg.JsonValue))
-		return "Error deserializing data", err
-	}
-
-	log.Println("getting weather for: ", weatherFuncArgs.Location)
-
-	output, err := getWeather(weatherFuncArgs.Location, s.Config.WeatherAPIKey)
-	if err != nil {
-		log.Println("Unable to get stock price: ", err)
-		return "There was a problem making that api call", err
-	}
-
-	return output, nil
-}
-
-func handleGetStockPrice(funcArg FuncArgs, apiKey string) (string, error) {
-	stockFuncArgs := StockFuncArgs{}
-	err := json.Unmarshal([]byte(funcArg.JsonValue), &stockFuncArgs)
-	if err != nil {
-		log.Println("Could not unmarshal func args: ", string(funcArg.JsonValue))
-		return "Error deserializing data", err
-	}
-
-	log.Println("getting price for: ", stockFuncArgs.Symbol)
-
-	output, err := getStockPrice(stockFuncArgs.Symbol, apiKey)
-	if err != nil {
-		log.Println("Unable to get stock price: ", err)
-		return "There was a problem making that api call", err
-	}
-
-	return output, nil
-}
-
-func handleMorningMessage(
-	ctx context.Context,
-	funcArg FuncArgs,
-	dgChannID string,
-	s *Skippy,
-) (string, error) {
-	morningMsgFuncArgs := MorningMsgFuncArgs{}
-	err := json.Unmarshal([]byte(funcArg.JsonValue), &morningMsgFuncArgs)
-	if err != nil {
-		log.Println("Could not unmarshal func args: ", string(funcArg.JsonValue))
-		return "Error deserializing data", err
-	}
-
-	output := toggleMorningMsg(
-		ctx,
-		morningMsgFuncArgs,
-		dgChannID,
-		s,
-	)
-
-	return output, nil
-}
-
-func getAndSendImage(
-	ctx context.Context,
-	funcArg FuncArgs,
-	channelID string,
-	s *Skippy,
-) (string, error) {
-	generateImageFuncArgs := GenerateImageFuncArgs{}
-	err := json.Unmarshal([]byte(funcArg.JsonValue), &generateImageFuncArgs)
-	if err != nil {
-		log.Println("Could not unmarshal func args: ", string(funcArg.JsonValue))
-		return "Unable to deserialize data", nil
-	}
-
-	imgUrl, err := GetImgUrl(generateImageFuncArgs.Prompt, s.AIClient)
-	if err != nil {
-		log.Println("unable to generate images", err)
-		return "Unable to generate image", err
-	}
-
-	log.Printf("recieved image url (%s) attempting to send on channel %s\n", imgUrl, channelID)
-	err = sendChunkedChannelMessage(s.DiscordSession, channelID, imgUrl)
-	return "image generated", err
-}
-
-func setReminder(
-	ctx context.Context,
-	funcArg FuncArgs,
-	channelID string,
-	s *Skippy,
-) (string, error) {
-	var channelMsg ReminderFuncArgs
-
-	err := json.Unmarshal([]byte(funcArg.JsonValue), &channelMsg)
-	if err != nil {
-		log.Println("Could not unmarshal func args: ", string(funcArg.JsonValue))
-		return "Unable to deserialize data", err
-	}
-
-	duration := time.Duration(channelMsg.TimerLength) * time.Second
-
-	log.Printf(
-		"attempting to send reminder on %s in %s\n",
-		channelID,
-		duration,
-	)
-
-	s.Scheduler.AddReminderJob(
-		channelID,
-		duration,
-		func() {
-			sendChunkedChannelMessage(s.DiscordSession, channelID, channelMsg.Message)
-			s.State.SetAwaitsResponse(channelID, true, s.AIClient)
-			for _, duration := range s.Config.ReminderDurations {
-				s.Scheduler.AddReminderJob(channelID, duration, func() {
-					sendAdditionalReminder(
-						ctx,
-						channelID,
-						channelMsg.UserID,
-						s,
-					)
-				})
-			}
-		},
-	)
-
-	return "worked", nil
-}
-
-func sendAdditionalReminder(
-	ctx context.Context,
-	channelID string,
-	userID string,
-	s *Skippy,
-) {
-	log.Println("sending another reminder")
-
-	// TODO: they should not get mentioned
-	if userID == "" {
-		userID = "they"
-	}
-
-	// TODO: maybe add this to additional_instruction
-	message := fmt.Sprintf(
-		"It looks %s haven't responsed to this reminder can you generate a response nagging them about it. This is not a tool request.",
-		UserMention(userID),
-	)
-
-	messageReq := openai.MessageRequest{
-		Role:    openai.ChatMessageRoleAssistant,
-		Content: message,
-	}
-
-	getAndSendResponseWithoutTools(
-		ctx,
-		channelID,
-		messageReq,
-		DEFAULT_INSTRUCTIONS,
-		s,
-	)
-}
-
-func UserMention(userID string) string {
-	if strings.HasPrefix(userID, "<@") && strings.HasSuffix(userID, ">") {
-		return userID
-	}
-	return fmt.Sprintf("<@%s>", userID)
-}
-
-func toggleMorningMsg(
-	ctx context.Context,
-	morningMsgFuncArgs MorningMsgFuncArgs,
-	channelID string,
-	s *Skippy,
-) string {
-	if !morningMsgFuncArgs.Enable {
-		s.Scheduler.CancelMorningMsgJob(channelID)
-		return "worked"
-	}
-
-	// TODO: this is hacky
-	// this should be fixed in the ai function definition
-	timeFmt := "15:04"
-	log.Println("Given time is: ", morningMsgFuncArgs.Time)
-	givenTime, err := time.Parse(timeFmt, morningMsgFuncArgs.Time)
-	if err != nil {
-		log.Printf("could not format time: %s. attempting to use AM/PM format", err)
-		timeFmt := "3:04 PM"
-		givenTime, err = time.Parse(timeFmt, morningMsgFuncArgs.Time)
-		if err != nil {
-			log.Printf("could not format time: %s.", err)
-			return "could not format time"
-		}
-
-	}
-
-	if s.Scheduler.HasMorningMsgJob(channelID) {
-		s.Scheduler.CancelMorningMsgJob(channelID)
-	}
-
-	log.Println("Setting the Morning Msg for: ", givenTime)
-
-	s.Scheduler.AddMorningMsgJob(
-		channelID,
-		givenTime,
-		func() {
-			sendMorningMsg(ctx, morningMsgFuncArgs, channelID, s)
-		},
-	)
-
-	return "worked"
-}
-
-func sendMorningMsg(
-	ctx context.Context,
-	morningMsgFuncArgs MorningMsgFuncArgs,
-	channelID string,
-	s *Skippy,
-) {
-	message := "Please tell everyone @here good morning."
-	for _, location := range morningMsgFuncArgs.WeatherLocations {
-		weather, err := getWeather(location, s.Config.WeatherAPIKey)
-		if err != nil {
-			log.Printf("unable to get weather for %s: %s\n", location, err)
-			continue
-		}
-		message += location + ":" + weather
-	}
-
-	for _, stock := range morningMsgFuncArgs.Stocks {
-		stockPrice, err := getStockPrice(stock, s.Config.StockAPIKey)
-		if err != nil {
-			log.Printf("unable to get weather for %s: %s\n", stock, err)
-			continue
-		}
-		message += stock + ":" + stockPrice
-	}
-
-	log.Println("getting morning message with prompt: ", message)
-
-	messageReq := openai.MessageRequest{
-		Role:    openai.ChatMessageRoleAssistant,
-		Content: message,
-	}
-
-	// reset the thread every morning
-	s.State.ResetOpenAIThread(channelID, s.AIClient)
-
-	getAndSendResponseWithoutTools(
-		ctx,
-		channelID,
-		messageReq,
-		MORNING_MESSAGE_INSTRUCTIONS,
-		s,
-	)
 }
 
 func submitToolOutputs(
