+++ skippy/tool_definitions.go
